refactor: Migrate EventDetailsFragment to use SafeArgs
Fixes #1198. Arguments have been added to the navigation graph XML file,
along with a more idiomatic Kotlin approach to creating argument bundles
for related fragments. The delegate function navArgs has been used to
retrieve the arguments for this fragment"
