+++ app/src/main/java/org/fossasia/openevent/general/ticket/TicketsFragment.kt
@@ -15,6 +15,7 @@ import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.navigation.Navigation.findNavController
+import androidx.navigation.fragment.navArgs
 import com.google.android.material.snackbar.Snackbar
 import org.fossasia.openevent.general.auth.SNACKBAR_MESSAGE
 import kotlinx.android.synthetic.main.fragment_tickets.ticketsCoordinatorLayout
@@ -27,6 +28,7 @@ import kotlinx.android.synthetic.main.fragment_tickets.view.ticketTableHeader
 import kotlinx.android.synthetic.main.fragment_tickets.view.ticketsRecycler
 import kotlinx.android.synthetic.main.fragment_tickets.view.time
 import org.fossasia.openevent.general.R
+import org.fossasia.openevent.general.event.EVENT_ID
 import org.fossasia.openevent.general.event.Event
 import org.fossasia.openevent.general.event.EventUtils
 import org.fossasia.openevent.general.utils.Utils.getAnimFade
@@ -35,27 +37,20 @@ import org.fossasia.openevent.general.utils.extensions.nonNull
 import org.fossasia.openevent.general.utils.nullToEmpty
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-const val EVENT_ID: String = "EVENT_ID"
-const val CURRENCY: String = "CURRENCY"
+const val CURRENCY: String = "currency"
 const val TICKET_ID_AND_QTY: String = "TICKET_ID_AND_QTY"
 
 class TicketsFragment : Fragment() {
     private val ticketsRecyclerAdapter: TicketsRecyclerAdapter = TicketsRecyclerAdapter()
     private val ticketsViewModel by viewModel<TicketsViewModel>()
-    private var id: Long = -1
-    private var currency: String? = null
+    private val safeArgs: TicketsFragmentArgs by navArgs()
     private lateinit var rootView: View
     private lateinit var linearLayoutManager: LinearLayoutManager
     private var ticketIdAndQty = ArrayList<Pair<Int, Int>>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        val bundle = this.arguments
-        if (bundle != null) {
-            id = bundle.getLong(EVENT_ID, -1)
-            currency = bundle.getString(CURRENCY, null)
-        }
-        ticketsRecyclerAdapter.setCurrency(currency)
+        ticketsRecyclerAdapter.setCurrency(safeArgs.currency)
 
         val ticketSelectedListener = object : TicketSelectedListener {
             override fun onSelected(ticketId: Int, quantity: Int) {
@@ -129,8 +124,8 @@ class TicketsFragment : Fragment() {
                 rootView.ticketInfoTextView.isGone = ticketTableVisible
             })
 
-        ticketsViewModel.loadEvent(id)
-        ticketsViewModel.loadTickets(id)
+        ticketsViewModel.loadEvent(safeArgs.eventId)
+        ticketsViewModel.loadTickets(safeArgs.eventId)
 
         return rootView
     }
@@ -146,7 +141,7 @@ class TicketsFragment : Fragment() {
 
     private fun redirectToAttendee() {
         val bundle = Bundle()
-        bundle.putLong(EVENT_ID, id)
+        bundle.putLong(EVENT_ID, safeArgs.eventId)
         bundle.putSerializable(TICKET_ID_AND_QTY, ticketIdAndQty)
         findNavController(rootView).navigate(R.id.attendeeFragment, bundle, getAnimSlide())
     }
